+++ backend/unithis/src/main/java/com/unithis/service/IImageService.java
@@ -9,7 +9,7 @@ import com.unithis.model.Image;
 public interface IImageService {
 
 	public int imageUpload(MultipartFile[] images, int id);
-	public int imageUpload(MultipartFile image, int id);
+	public int imageUpload(MultipartFile image, long id);
 	public List<Image> getImage(int id);
 	
 }
